Remove explicit content-length header for GET requests (#3278)
* Remove explicit content-length header workaround which no longer seems necessary

Fixes #3277

Originally this was introduced in #1189 to fix https://github.com/aws/aws-sdk-php/issues/701

In my case however, I receive a 500 when trying to stream content from
media.licdn.com because of this header.

I then went back and tried the example from the aws-sdk-php and removed
that header only to learn it works as expected.

My conclusion therefore is that in the past 10 years either AWS server
or the SDK changed and this workaround, once needed, no longer is
necessary and we can remove it.

* Keep explicit content-length: 0 but only for PUT & GET requests

Like `CurlHandler` in `\GuzzleHttp\Handler\CurlFactory::applyMethod`
currently does it.

Fixes #3277